ref : move route recommend type to service recommend type file
