+++ src/services/recipes/recipe/helper.ts
@@ -8,7 +8,7 @@ import {
 } from "./type.js";
 import { config } from "../../../config/index.js";
 import { sanitizeRecipeData, generateRecipeKey } from "../helper.js";
-import { RecommendRecipe } from "../../../router/v1/recommend/type.js";
+import { RecommendRecipe } from "../../recommends/type.js";
 import { decrypt } from "../../../utils/encrypt.js";
 import { shuffleArray } from "../../../utils/shuffle.js";
