+++ src/services/recipes/recipe/recipeRecommendService.ts
@@ -2,7 +2,7 @@ import { RowDataPacket } from "mysql2";
 
 import mysqlDB from "../../../db/mysql.js";
 import { decryptRecipeURLAndGetRecipeId } from "./helper.js";
-import { RecommendRecipe } from "../../../router/v1/recommend/type.js";
+import { RecommendRecipe } from "../../recommends/type.js";
 import { validateId } from "../../../utils/numbers.js";
 import { arrayToPlaceholders } from "../../../utils/query.js";
 import { ServiceError } from "../../helpers/ServiceError.js";
